requirements.txt
requests==2.23.0
prometheus_client==0.7.1
pandas==1.0.4
PyYAML==5.3
click==7.1.2
progressbar2==3.51.3
timeloop==1.0.2
rope==0.17.0
autopep8==1.5.3
twine==3.1.1